requirements.txt
# Abhängigkeiten für den Museum Guide Chatbot
# Installiere mit: pip install -r requirements.txt

# Rasa Framework - Hauptabhängigkeiten
rasa==3.6.13
rasa-sdk==3.6.2

# HTTP Requests für API-Aufrufe
requests==2.31.0

# Umgebungsvariablen-Management
python-dotenv==1.0.0

# Streamlit für Web-Interface
streamlit==1.28.0

# Optionale Abhängigkeiten für erweiterte Funktionalität

# Für bessere NLP-Funktionen (optional)
spacy==3.7.2

# Für Datenbank-Verbindungen (optional, für Produktionsumgebung)
# psycopg2-binary==2.9.9  # PostgreSQL
# pymongo==4.6.0          # MongoDB

# Für erweiterte Logging-Funktionen (optional)
# structlog==23.2.0

# Für Web-Interface (optional)
# flask==3.0.0
# flask-cors==4.0.0

# Für erweiterte Testing-Funktionen (optional)
# pytest==7.4.3
# pytest-asyncio==0.21.1

# Für Performance-Monitoring (optional)
# prometheus-client==0.19.0

# Für Deployment (optional)
# gunicorn==21.2.0
